+++ flutter_appauth/lib/flutter_appauth.dart
@@ -7,7 +7,7 @@ export 'package:flutter_appauth_platform_interface/flutter_appauth_platform_inte
         AuthorizationTokenResponse,
         EndSessionRequest,
         EndSessionResponse,
-        ExternalAgentType,
+        ExternalUserAgent,
         FlutterAppAuthOAuthError,
         FlutterAppAuthPlatformErrorDetails,
         FlutterAppAuthUserCancelledException,