+++ flutter_appauth_platform_interface/lib/src/authorization_token_request.dart
@@ -1,6 +1,6 @@
 import 'authorization_parameters.dart';
-import 'external_agent_type.dart';
-import 'grant_types.dart';
+import 'external_user_agent.dart';
+import 'grant_type.dart';
 import 'token_request.dart';
 
 /// Details required for a combined authorization and code exchange request
@@ -18,8 +18,8 @@ class AuthorizationTokenRequest extends TokenRequest
     super.discoveryUrl,
     List<String>? promptValues,
     super.allowInsecureConnections,
-    ExternalAgentType preferredExternalAgent =
-        ExternalAgentType.asWebAuthenticationSession,
+    ExternalUserAgent externalUserAgent =
+        ExternalUserAgent.asWebAuthenticationSession,
     super.nonce,
     String? responseMode,
   }) : super(
@@ -27,7 +27,7 @@ class AuthorizationTokenRequest extends TokenRequest
         ) {
     this.loginHint = loginHint;
     this.promptValues = promptValues;
-    this.preferredExternalAgent = preferredExternalAgent;
+    this.externalUserAgent = externalUserAgent;
     this.responseMode = responseMode;
   }
 }