+++ flutter_appauth_platform_interface/lib/src/end_session_request.dart
@@ -8,7 +8,7 @@ class EndSessionRequest with AcceptedAuthorizationServiceConfigurationDetails {
     this.postLogoutRedirectUrl,
     this.state,
     this.allowInsecureConnections = false,
-    this.preferredExternalAgent = ExternalAgentType.asWebAuthenticationSession,
+    this.externalUserAgent = ExternalUserAgent.asWebAuthenticationSession,
     this.additionalParameters,
     String? issuer,
     String? discoveryUrl,
@@ -38,17 +38,9 @@ class EndSessionRequest with AcceptedAuthorizationServiceConfigurationDetails {
   /// This property is only applicable to Android.
   bool allowInsecureConnections;
 
-  /// Decides what type of external agent to use for the authorization flow.
-  /// ASWebAuthenticationSession is the default for iOS 12 and above.
-  /// EphemeralSession is not sharing browser data
-  /// with the user's normal browser session but not keeping the cache
-  /// SFSafariViewController is not sharing browser data
-  /// with the user's normal browser session but keeping the cache.
-  /// This property is only applicable to iOS versions 13 and above.
-  /// ExternalAgentType? preferredExternalAgent;
-  ///
-  /// Sign in and out must have the same type.
-  ExternalAgentType? preferredExternalAgent;
+  /// Specifies the external user-agent to use.
+  ExternalUserAgent? externalUserAgent;
 
+  /// Additional parameters to include in the request.
   final Map<String, String>? additionalParameters;
 }